staging-tests/requirements.txt
# Comprehensive Integration Test Requirements
# This file includes all dependencies for the integration test suite

# Core Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
pytest-timeout==2.2.0
pytest-html==4.1.1
pytest-xdist==3.5.0  # Parallel test execution

# HTTP Client and API Testing
requests==2.31.0
urllib3==2.1.0

# AWS SDK and Services
boto3==1.34.0
botocore==1.34.0

# Database Integration Tests
psycopg2-binary==2.9.7
sqlalchemy==2.0.23

# Cache Integration Tests  
redis==5.0.1
hiredis==2.3.2  # Performance improvement for redis

# Messaging Integration Tests
kafka-python==2.0.2
confluent-kafka==2.3.0  # Alternative Kafka client

# Test Data Generation
faker==22.0.0
factory-boy==3.3.0

# Performance Testing
locust==2.20.0  # Alternative to K6 for Python-based load testing

# Security Testing
python-owasp-zap-v2.4==0.0.22
bandit==1.7.6

# Monitoring and Observability
prometheus-client==0.19.0

# Configuration Management
python-dotenv==1.0.0
pyyaml==6.0.1

# Data Validation and Serialization
pydantic==2.5.3
dataclasses-json==0.6.1
marshmallow==3.20.1

# Utilities
typing-extensions==4.8.0
python-dateutil==2.8.2
pytz==2023.3.post1

# Logging and Reporting
colorlog==6.8.0
tabulate==0.9.0
jinja2==3.1.2

# Code Quality (for test code)
black==23.12.1
flake8==7.0.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0